requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
redis==5.0.1
python-socketio==5.10.0
eventlet==0.33.3
WTForms==3.1.1
python-dotenv==1.0.0
bcrypt==4.1.2